// pronti-frontend-main/front/src/app/modulos/cargo-vacacion/form/form-cargo-vacacion.enum.ts
export enum FormCargoVacacionEnum {
  // nombre = 'nombre',
  fechaDesde = 'fechaDesde',
  fechaHasta = 'fechaHasta',
  diasVacaciones = 'diasVacaciones',
  diasAntiguedad = 'diasAntiguedad',
  diasTeoricos = 'diasTeoricos',
  diasTomados = 'diasTomados',
  diasSaldo = 'diasSaldo',
  valorVacaciones = 'valorVacacion',
  valorDias = 'valorDias',
  valorAntiguedad = 'valorAntiguedad',
  valorTeorico = 'valorTeorico',
  valorTomado = 'valorTomado',
  valorSaldo = 'valorSaldo',
  totalIngresosAnio = 'totalIngresosAnio',
  numAnioAcumulado = 'numAnioAcumulado',
  idTrabajdor = 'idTrabajdor',
}